Create resource instance with plain php
Do this with plain php instead of the IoC container of Laravel. In Laravel 5.4 it's not possible to call app()->make() with dependency injection.
So a new empty client will be created and a 'malformed url error will be thrown'.
That nasty bugger took me a while to figure out.